requirements.txt
fastapi
uvicorn
openai
chromadb
tiktoken
python-dotenv
discord.py
requests
